Removes Ingress, uses service as load balancer/ingress.
This change allows us to connect to the web frontend, but it's not
possible to render prov files or add them, because of CORS errors.
The reason for this is that with this type of ingress we do not
know the hostname under which the service will run after deploying.
However, we need to know the hostname before deploying to pass it
into the environment of the container.

TODO: refactor the docker image to use mounted volumes for the secrets
and rewrite run-container to create the self signed certs and make them
available to the running container in a volume. Do the same for a
configuration file that is used by app.py. Do the same for the wsgi
configuration file and write a wrapper that is used as the container
command to append the servername to httpd.conf, after checking if this
is still necessary.